pyproject.toml
[tool.poetry]
name = "pool-ticket-task"
version = "0.1.0"
description = ""
authors = ["mr.khavari <mr.khavari.n@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic-settings = "^2.6.1"
black = "^24.10.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.19"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
aioredis = "^2.0.1"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
